updated MessageTemplate name, moved mockedData, all types in one file
